+++ R/hmpMDPWriter.R
@@ -42,7 +42,7 @@
 #' @param getLog Output log text.
 #' @return A list of functions.
 #' @note Note all indexes are starting from zero (C/C++ style).
-#' @example inst/examples/hmpMDPWriter.R
+#' @example inst/examples/hmpMDPWriter-ex.R
 #' @export
 hmpMDPWriter<-function(file="r.hmp", rate=0.1, rateBase=1, precision=0.00001, desc="HMP file created using hmpMDPWriter in R", getLog = TRUE) {
 	# addLevelRates<-function(rates){