+++ NEWS.md
@@ -1,3 +1,8 @@
+# MDP2 v2.1.2
+
+* Modified MDP object to class HMDP
+* Added a plot method for class HMDP
+
 # MDP2 v2.1.1
 
 * Fixed FORTRAN error in R-devel